Move deployments and stage management out of project
Instead, we print the `aws` command to use to deploy in a stage.

Terraform does not support well the API Gateway deployment and stage
resources.  Triggering the deployment upon relevant IaC changes and
deploying at the right moment in a `terraform apply` sequence is too
complex.

Also, it makes sense to deploy in production only through a manual
action, or after running some tests in a CD pipeline.